requirements.txt
# Trading signals system dependencies
aiohttp==3.9.1
websockets==12.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1

# Type checking
mypy==1.8.0
types-requests==2.31.0

# Linting
flake8==6.1.0
black==23.12.1